Sound_Localization.py and mytest_avnet.py were deleted and mytest_avnet_video.py was added.
